requirements.txt
yowsup2
promise
python-axolotl
Pillow
